Karel code review diff comments for AppInitSDK
