+++ frontend/config/dev.webpack.config.js
@@ -59,15 +59,15 @@ const webpackProxy = {
     },
     {
       // if you want to host different plugin than `console-demo-plugin` locally adjust this line
-      context: ['/beta/api/plugins/hac-dev/', '/api/plugins/hac-dev/'],
+      context: ['/beta/api/plugins/hac-dev', '/api/plugins/hac-dev'],
       // In order to serve your plugin locally on your server change this line ↓
       target: 'http://localhost:8003',
       secure: false,
       changeOrigin: true,
-      pathRewrite: {
-        // if you don't want to rewrite `/beta/api/plugins` to `/api/plugins` remove this line
-        // '^/beta/api/plugins/console-demo-plugin': '/api/plugins/console-demo-plugin',
-      },
+      // pathRewrite: {
+      //   // if you don't want to rewrite `/beta/api/plugins` to `/api/plugins` remove this line
+      //   '^/beta/api/plugins/console-demo-plugin': '/api/plugins/console-demo-plugin',
+      // },
     },
   ],
 };
