+++ frontend/config/plugins.js
@@ -16,7 +16,10 @@ const plugins = [
     },
     shared: [
       { 'react-router-dom': { singleton: true, requiredVersion: package.dependencies['react-router-dom'] } },
+      { 'react-redux': { singleton: true, requiredVersion: package.dependencies['react-redux'] } },
       { '@openshift/dynamic-plugin-sdk-utils': { singleton: true, requiredVersion: package.dependencies['@openshift/dynamic-plugin-sdk-utils'] } },
+      { '@openshift/dynamic-plugin-sdk': { singleton: true, requiredVersion: package.dependencies['@openshift/dynamic-plugin-sdk'] } },
+      { 'Sdk/createStore': { singleton: true } }
     ],
   }),
   new CopyWebpackPlugin({
