+++ frontend/package.json
@@ -52,7 +52,7 @@
     "react-dom": "17.0.2",
     "react-helmet": "^6.1.0",
     "react-i18next": "^11.14.2",
-    "react-redux": "7.2.4",
+    "react-redux": "7.2.6",
     "react-router": "^5.2.0",
     "react-router-dom": "5.2.0",
     "read-pkg": "^5.2.0",
