+++ frontend/src/poc-code/console-dynamic-plugin-sdk/src/shared-modules-init.ts
@@ -18,6 +18,7 @@ const modules: SharedModuleResolution = {
   redux: async () => () => require('redux'),
   'react-helmet': async () => () => require('react-helmet'),
   'react-i18next': async () => () => require('react-i18next'),
+  'react-redux': async () => () => require('react-redux'),
   'react-router': async () => () => require('react-router'),
   'react-router-dom': async () => () => require('react-router-dom'),
 };
