+++ frontend/src/store/index.ts
@@ -1,6 +1,7 @@
 import { getRegistry } from '@redhat-cloud-services/frontend-components-utilities/Registry';
 import promiseMiddleware from 'redux-promise-middleware';
 import { notificationsMiddleware } from '@redhat-cloud-services/frontend-components-notifications/notificationsMiddleware';
+import { SDKReducers } from '@openshift/dynamic-plugin-sdk-utils';
 
 let registry;
 
@@ -10,6 +11,7 @@ export function init(...middleware) {
     notificationsMiddleware({ errorDescriptionKey: ['detail', 'stack'] }),
     ...middleware.filter(Boolean),
   ]);
+  registry.register(SDKReducers);
   return registry;
 }
