+++ frontend/tsconfig.json
@@ -17,7 +17,9 @@
       "@console/plugin-sdk": ["src/poc-code/console-plugin-sdk"],
       "@console/plugin-sdk/*": ["src/poc-code/console-plugin-sdk/*"],
       "@console/dynamic-plugin-sdk": ["src/poc-code/console-dynamic-plugin-sdk"],
-      "@console/dynamic-plugin-sdk/*": ["src/poc-code/console-dynamic-plugin-sdk/*"]
+      "@console/dynamic-plugin-sdk/*": ["src/poc-code/console-dynamic-plugin-sdk/*"],
+      "Sdk": ["src/sdk"],
+      "Sdk/*": ["src/sdk/*"]
     },
     "typeRoots": ["node_modules/@types", "@types"],
     "types": ["node", "jest"],
